refactor cfg unit tests
Split cfg_test.c into multiple files based on the function
under test to reduce file size and complexity.
Also replaced usage of strcmp with streq from own library.
